refactor: rename METCalculator to METProvider
The METCalculator class was renamed to METProvider to better reflect its function of providing MET values. This change was also propagated to the test suite.

The StationarySegmentMergerTest was deleted as it is no longer needed.

PolylineFormatter constructor was updated to accept a Logger instance. This change was reflected in its test classes, PolylineProcessorTest and PolylineStatsWorkoutLogItemFormatterTest, by providing a TimberAppLogger instance during formatter initialization.